Ensure sockets don't get prematurely GC'd when referenced by a protocol
This is a limitation in how Lua and C++ interact. Even though the
socket references the protocol in C++, Lua has no way of knowing this.
As a result, it would GC the socket. This exposed several bugs and
unexpected types of failure, which was very elucidating. The core issue
of exposing the reference isn't automatic, as I would have preferred,
but it's reasonable enough.